+++ frontend/lost/src/api/dataset_export_api.tsx
@@ -2,7 +2,7 @@ import { saveAs } from 'file-saver'
 import { useMutation, useQuery, useQueryClient } from 'react-query'
 import { showError, showSuccess } from '../components/Notification'
 import { httpClient } from '../actions/http-client'
-import { AvailableDatasetExportsResponse } from './model'
+import { AvailableDatasetExportsResponse } from '../types/model'
 
 export const useExportDataset = () => {
   const queryClient = useQueryClient()
