+++ frontend/lost/src/api/statistics_api.jsx
@@ -1,5 +1,5 @@
 import axios from 'axios'
-import { API_URL } from '../../lost_settings'
+import { API_URL } from '../lost_settings'
 import { useMutation } from 'react-query'
 
 export const usePersonalStatistics = () => {
